backend/requirements.txt
agno[os]>=2.5.0
openai>=1.0.0
anthropic>=0.40.0
python-dotenv>=1.0.0
pymupdf>=1.27.0
lancedb>=0.30.0
pypdf>=6.0.0
fastembed>=0.8.0
pandas>=2.0.0
duckdb>=1.0.0
python-pptx>=1.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
